Formally allow serpent form to wear helmets
This is at least as much a stopgap bugfix as it is a deliberate design
change, but this fixes a couple existing issues that are tricky to resolve
otherwise.

(Currently if you're wearing a helmet before entering the form, it will
stay on, but you aren't allowed to equip one *after* entering the form. But
actually preventing the player from transforming if they have a helmet on
has some awkwardnesses and we can't just remove the helmet automatically
(or this would shatter cursed equipment!) and selectively melding some but
not all headgear would likely be a source of new bugs after feature freeze,
so this seems like the simplest solution for the moment.)

(It's not really a balance issue for the form to wear helmets,
incidentally; some people just think the flavor is odd.)