Fix console sometimes not printing a second line of statuses
CPRINTF would automatically wrap the cursor's column back to the start of
the HUD region if printing a status that reached the end of available
space, causing _print_status_lights() to falsely think there was additional
space available on the current line (resulting in it not rendering the
second line at all).

Since it effectively does its own word-wrapping (and, perhaps unsafely,
relies on the cursor's current position to do so), it seems best to print
without wrapping.

This fixes #4502