Improve badmut messages for melded gear
As pointed out by the previous commit, the code to summarize a badmut for
what gear is melded by a form was slightly incorrect. In the process of
fixing this, I have also improved it by being more specific (if room
allows) about what is melded - accounting for whether the player had a slot
of that type to meld in the first place.