Pre-reveal Gauntlet floors (aside from the minotaur area)
This has been a long-requested QoL improvement, which was only deferred
due to not wanting to pre-reveal the minotaur's area (which is *intended*
to remain secret.) But now the no_automap tag can be added to vaults,
which prevents them from being revealed by brflag::fully_map. This is used
here to mask off the minotaur's subvaults, while revealing the rest of
Gauntlet.