New mutation: doubled potion durations
As a much-delayed follow-up to 9235b1e, 9303715, and a98c639, this commit
kicks off another batch of new positive mutations. These are meant to add
more interest and more high-roll possibilities beyond resists or flat
stats, which should hopefully make mutation roulette more varied and
tempting, as well as make up for the inherent lack of control one has over
curing bad mutation currently via potions of mutation.

This rare mutation repurposes some of the old functionality of armataur
tongues as was simplified and refocused in afeb8ed towards healing potions
for Oni. (It's flavoured as "efficient metabolism" rather than a long
tongue mostly for the sake of formicids and shapeshifting.) Much like said
focus, this mutation focuses the effect to be very specific instead of
needing explicit clarifications or categorizations, in affecting the
majority of potions (if not the strategic or healing ones) by doubling the
duration of buffs (and debuffs) they grant. (There is a bit of a shame in
not supporting even further high-rolls with potions of experience, but the
simplicity of this mutation compared to the special cases across old
armataur tongues hopefully feels worth the cost of losing such). With
essentially every character benefitting from using potion buffs before
engaging with uniques or branch end vaults, this should be reasonably
valuable and distinct without being particularly complex or burdensome,
while also still being reasonably capped in power by the number of potions
that drop per game.

(The implementation of this also adjusts several status caps for being
relatively arbitrary- considering ambrosia, attraction, enlightenment,
lignification, and resistance don't have caps, I'm not sure why might and
brilliance need caps on their effects either. This also increases the
cap on haste from 40 to 80, with potions of haste properly outlined as
giving 26-40 deca-auts of haste instead of a previous 40-79 being changed
to 26-52 followed by a cap of 40 due to the strange handling a long whiles
ago in 13d70d82c4. These alltogether are a little bit more power to Gozag
Potion Petition (which has gotten an adjusted description to suit the
various interactions it doesn't actually have), which can be corrected in
other ways if needed, but being more consistent across more potion effects
is a better position to be starting from.)

(The icon for this mutation modifies ontoclasm's standard potion tiles.)