Fix hellfire mortar being able to dig through temp terrain (Back-form)
When an effect digs through or otherwise destroys temporary terrain, it is
supposed to revert back to the original terrain (unless that original
terrain could be destroyed by the same means). This is to prevent, for
instance, turning a metal wall into a tree with Summon Forest and then
digging through that tree.

But due to some oddities in Hellfire Mortar's own digging code (and the
fact that digging *tracers* didn't follow the above rule), it *could* dig
through any sort of wall which had been turned into a tree. Now, hopefully,
it will not.