Increase stun immunity duration post-Vex
A variety of people have been severely wrecked by sphinx marauders since
their addition, and while this has been good for punishing those ignoring
their central gimmicks + raising the value of Will in Shoals as
differentiates it from Swamp (though people still undervalue it versus
fenstriders, strangely enough). It is, however, admittedly quite
frustrating to foreground more stuns in the game, especially on a monster
that does some of the most reliable high damage in the game through
AF_AIRSTRIKE.

The stun immunity duration granted after Vex was set to the same duration
as the one after Paralysis (1 to 3 turns), and to reduce the rather severe
low-rolls a player could experience in being Vexed again immediately
afterwards (and to match how it's longer than Paralysis itself), I am
instead setting it to the same duration granted after Sleep (3 to 5 turns).
This should still leave marauders plenty dangerous overall with the first
Vex plus their AF_AIRSTRIKE damage, but prevent the worst scenarios of
being vexed, drinking enlightenment or putting on a Will+ ring, and then
being vexed again while the sphinx is much closer to result in a rapid
death for frailer characters.

(This also technically affects Dominate Undead for undead player
characters, but the spell's already quite nasty as is even in its current
limited usage due to the full screen effect, and bloodprinces have Grave
Claw to terrify the non-poltergeists with.)