Somewhat standardize and simplify a few more god wraths
As elaborated upon with the last god wrath change of 7aa09ac, many god
wraths are antiquated designs that are easily rested out. With the
improvements to tension tracking, I'm deploying a standardized tension
check (plus a sanity clause of an HP check) to reduce the effectiveness
of resting out a variety of god effects, trying to preserve distinct
effect flavour + reducing waiting out wraths + focusing long lists
towards a simpler or more concentrated pool in general.

 * Fedhas wrath no longer deploys direct damage (either boringly
   irrelevant or extremely dissatisfying in its results) via spells or
   giant spore and instead checks tension to apply Fungus / Tree
   transformations (both highly appropriate and relatively rare in hostile
   circumstances). The oklob spawning effect also no longer preserves the
   very old and very complicated Reproduction code to instead just place
   fuller rings of briars around the player alongside oklobs in a loose
   2-to-5 space radius around the player.

 * Jiyva wrath no longer transforms the player; Jiyva's wrath is already
   so heavily focused on strategic penalties that such wrath doesn't really
   need transformations distinctly not very slimy in form.

 * Kikubaaqudgha wrath no longer randomly torments (as is no longer Kiku's
   big invocation nor much easy to make satisfyingly relevant) and no
   longer checks rN to be thus fully resistable. The wrath now just uses
   the new tension check to either apply a short xl-scaled Sign of Ruin
   effect or drain for somewhat less than previously.

 * All gods besides the good gods, instead of currently just Lugonu, no
   longer permit rejoining them if one is already under their wrath. This
   prevents some strange exploits like abandoning Jiyva to kill TRJ and
   grab Slime:$ rewards without active item-eating slimes, then rejoining
   Jiyva who is otherwise kept alive during wrath. (The good gods already
   have weird wrath rules as is, so this shouldn't be too much of an
   outlier, and they'd never permit rejoining under such rules without
   deliberately joining a forbidden god to actively get their wrath.)

This is only a small part of future total god wrath plans- with more
passive effects, some reliance on potential new mechanics, more tension
and space checks to avoid resting out wrath, and more- but it deals with
a handful of more obviously exploitable matters, if nothing else.