Add Bane of Lethargy
A simple one to start with - increased move delay by 30% while active.

Slow movement was a badmut people hated to be stuck with back in the days
when it was in the random mutation pool, but I hope the context of being
able to work it off in a moderate length of time will make it a more
interesting penalty to deal with.