Add Bane of the Reckless
This sets the player's SH to 0 (and lasts a relatively long time). Since
this obviously does nothing to some players, it will only be given to a
player who currently has non-temporary SH (which does include amulets of
reflection or Qazlal worship). Since it can be 'mitigated' by different
equipment setups, the hope is that the duration will sometimes result in
players temporarily choosing different item sets than they would otherwise
(such as using a two-hander for a while).

It only being given to players who have SH at the time the bane is applied
*technically* makes it possible to game this, but I'm not convinced this
is actually beneficial (such as taking off a shield you're actually using,
while fighting a doom user, so that you... won't have a tiny chance of
losing the SH bonus that you just voluntarily lost at the cost of 5 turns
sitting still and instead get something possibly worse?)