Fix a (harmless) fallthrough in _on_enter_form (hyperactiveChipmunk)
Entering treeform fell through to entering dragon because of a shared
net-escaping effect. But when this intentional fallthrough was first
written, entering dragon form did nothing else. Now it does something
dragon-specific in addition. By coincidence, it is harmless to do this
when entering tree form, but it seems possible to cause a subtle bug if
that ever changes in the future, so I'm changing it now anyway.