Give Vehumet's first gift on join (Kab)
Vehumet's always been a bit slow to offer benefits, and there's been debate
over raising the starting level of spell gifts to 2 or 3 so that the first
couple are more likely to be useful (see 602a2ca for one step in this
direction). However, we do try to make most gods useful for a wide variety
of starts, and raising the level of spell gifts might make Vehumet even less
appealing for characters with questionable int or uh, melee deep elves (?).

So instead, let's let Vehumet gift the level 1 spell immediately upon joining
to make it a bit more tempting. The timing of other Vehumet gifts is unchanged
for now. Historically, gifts on joining have been considered "too strong" for
most gods, but access to a single level 1 spell should be fine.