Fix bat swarm form getting torment immunity (Acrobat)
While it is likely fine from a balance perspective if your cooldown ability
gave you rTorment, bat swarm is already doing quite a lot of different
things and it's not at all clear or expected this would be one of them, so
I am removing it. (It was only ever true by accident).