Add Bane of Multiplicity
This bane periodically creates 2-3 fragile clones of a monster in the
player's sight. These clones have 1/3rd the health of the original, 1/2 the
spell HD, and do 2/3rd attack damage, but do not expire on their own. The
interval at which they are created is heavily randomized, but it must be
at least 15 turns since the last time it happened (and usually a fair bit
more than that).

Uniques are not eligible to be cloned (I felt it might be too big and
outlier of threat).

---------------

Figment status icon made by regret-index.