Allow undead hydra to attack multiple targets
For some reason, undead hydra were not able to perform the 'attack multiple
targets in one round if the first target dies' thing that living hydra do.
Maybe the original argument was that zombies weren't intelligent enough?
But zombies are still perfectly capable of using complex flight muscles or
coiling themselves around enemies to constrict them, so I don't see why
this should be any different (also: I'm not sure anyone even realized this
was a thing in the first place.)