Simplify Serpent of Hell melee
The Serpent of Hell used a system like hydras to multiply its first attack
but the number of heads it had - but would stictly and always have exactly
3. This technically allowed it to have slightly more than 4 attacks per
round and gave it hydra-like pseudo-cleave against dead enemies, but
neither of these seem terribly relevant (the '5th' attack only did 15
damage!). It also made /?m lookup lie by reporting its damage as if it had
only 1 head.

This feels like needless code complexity for an effect that doesn't
even interact with the head system dynamically, so I've just baked 3 bite
attacks into its attack table in the usual way, and slightly buffed the
remaining aux attacks in compensation for losing one. It should be hard to
even notice this change in a real game.