Fix a lingering chance for Zot:5 to be double-trapped (knorpule3000)
When 19588f4 tried to make it so both main entrances to the Zot:5 vault
can't be cut off by traps, it neglected to use different floor glyphs
than the rest of the vault, and thus accidentally left a very tiny
chance for the vault's completely random trap placement on all floor tiles
afterwards to choose spaces that weren't supposed to have traps. This
should fix such.